+++ core/src/com/luzi82/shinju/ShinjuCommon.java
@@ -20,11 +20,11 @@ public class ShinjuCommon {
 	public World.Data mShinjuData;
 	public World.Model mShinjuModel;
 
-	public static final int CELL_SIZE = 256;
+	public static final long CELL_SIZE = 256;
 
-	public static final int HERO_SIZE = 1 * CELL_SIZE;
-	public static final int WITCH_SIZE = 2 * CELL_SIZE;
-	public static final int WITCHSEED_SIZE = 1 * CELL_SIZE;
+	public static final long HERO_SIZE = 1 * CELL_SIZE;
+	public static final long WITCH_SIZE = 2 * CELL_SIZE;
+	public static final long WITCHSEED_SIZE = 1 * CELL_SIZE;
 
 	public static final float PHI = (1f + (float) Math.sqrt(5)) / 2f;
 	public static final float PHI2 = 1f + PHI;
